@@ -1,8 +1,42 @@
 {
   "name": "cyberspectrum/php-transifex",
+  "description": "transifex API client written 100% in PHP",
   "license": "MIT",
   "type": "project",
-  "description": "transifex API client written 100% in PHP",
+  "require": {
+    "php": "^8.1",
+    "ext-json": "*",
+    "ext-mbstring": "*",
+    "jane-php/json-schema-runtime": "^7.5.0",
+    "jane-php/open-api-runtime": "^7.5.0",
+    "php-http/cache-plugin": "^1.7.5",
+    "php-http/client-common": "^2.6.1",
+    "php-http/discovery": "^1.15",
+    "php-http/httplug": "^2.4.0",
+    "php-http/message": "^1.14.0",
+    "php-http/message-factory": "^1.1.0",
+    "php-http/multipart-stream-builder": "^1.2.0",
+    "php-http/promise": "^1.1.0",
+    "psr/cache": "^3.0",
+    "psr/http-client": "^1.0",
+    "psr/http-factory": "^1.0",
+    "psr/http-message": "^1.1",
+    "psr/log": "^3.0.0",
+    "symfony/deprecation-contracts": "^3.2",
+    "symfony/options-resolver": "^6.2",
+    "symfony/serializer": "^6.2",
+    "symfony/validator": "^6.2"
+  },
+  "require-dev": {
+    "http-interop/http-factory-guzzle": "^1.2.0",
+    "php-http/guzzle7-adapter": "^1.0.0",
+    "phpcq/runner-bootstrap": "^1.0@dev"
+  },
+  "suggest": {
+    "php-http/buzz-adapter": "For using buzz as transport",
+    "php-http/guzzle7-adapter": "For using guzzle as transport",
+    "php-http/react-adapter": "For using react as transport"
+  },
   "autoload": {
     "psr-4": {
       "CyberSpectrum\\PhpTransifex\\": "src"
@@ -13,28 +47,15 @@
       "CyberSpectrum\\PhpTransifex\\Tests\\": "tests"
     }
   },
-  "require": {
-    "php": "^5.6 | ^7.0",
-    "php-http/cache-plugin": "^1.2",
-    "php-http/client-common": "^1.3",
-    "php-http/discovery": "^1.1",
-    "php-http/httplug": "^1.1",
-    "psr/cache": "~1.0"
-  },
-  "require-dev": {
-    "phpcq/all-tasks": "~1.2",
-    "php-http/guzzle6-adapter": "~1.1"
-  },
-  "suggest": {
-    "php-http/buzz-adapter": "For using buzz as transport",
-    "php-http/guzzle5-adapter": "For using guzzle 5 as transport",
-    "php-http/guzzle6-adapter": "For using guzzle 6 as transport",
-    "php-http/react-adapter": "For using react as transport"
+  "config": {
+    "allow-plugins": {
+      "php-http/discovery": false
+    },
+    "sort-packages": true
   },
   "extra": {
     "branch-alias": {
-      "dev-master": "1.0.x-dev",
-      "dev-develop": "1.1.x-dev"
+      "dev-master": "2.0.x-dev"
     }
   }
 }
